Rename useFocusItemsHelper --> useScrollIToItem
