+++ generatedTypes/hooks/useScrollIToItem/index.d.ts
@@ -52,5 +52,5 @@ export declare type ResultProps = {
      */
     focusIndex: (index: number, animated?: boolean) => void;
 };
-declare const useFocusItemsHelper: (props: Props) => ResultProps;
-export default useFocusItemsHelper;
+declare const useScrollIToItem: (props: Props) => ResultProps;
+export default useScrollIToItem;
