+++ src/hooks/index.ts
@@ -2,4 +2,4 @@ export {default as useToggleValue} from './useToggleValue';
 export {default as useDidUpdate} from './useDidUpdate';
 export {default as useScrollEnabler} from './useScrollEnabler';
 export {default as useScrollReached} from './useScrollReached';
-export {default as useFocusItemsHelper} from './useFocusItemsHelper';
+export {default as useScrollIToItem} from './useScrollIToItem';
